style: fix ruff and ty diagnostics
- Remove unused import/variable in database.py
- Use contextlib.suppress in client.py (SIM105)
- Change QDRANT_TIMEOUT type from float to int for qdrant-client stubs
- Add ty ignore comments for langchain_openai ChatOpenAI kwargs in rewriter.py
- Fix import sorting, whitespace, and f-string warnings across scripts/tests

+++ app/core/config.py
@@ -105,7 +105,7 @@ class Settings(BaseSettings):
     QDRANT_URL: str = "http://localhost:6333"
     QDRANT_API_KEY: str | None = None
     QDRANT_COLLECTION_NAME: str = "knowledge_chunks"
-    QDRANT_TIMEOUT: float = 10.0
+    QDRANT_TIMEOUT: int = 10
     QDRANT_RETRIES: int = 3
 
     # Reranker / Rewriter
